+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.2.0 - 2018-03-09
+* Added new module: getRGB
+* Added new module: parseValue (postcss-value-parser)
+* Added new module: sassFunction
+
 # 2.1.1 - 2017-10-11
 * Fixing some issues with new modules